Fixed minor warning.

+++ ofstd/libsrc/ofwhere.c
@@ -202,7 +202,7 @@ int OFgetExecutablePath(char* out, int capacity, int* dirname_length)
     length = (int)strlen(resolved);
     if (length <= capacity)
     {
-      memcpy(out, resolved, length);
+      memcpy(out, resolved, (size_t)length);
 
       if (dirname_length)
       {
